move gpu specialization to a separate file
